requirements.txt
pandas
numpy
scikit-learn
joblib
dvc
mlflow
matplotlib
seaborn
jupyter
pyyaml
